+++ ALSpringBackend/src/main/java/com/ardaslegends/data/service/dto/army/UpkeepDto.java
@@ -1,4 +1,4 @@
 package com.ardaslegends.data.service.dto.army;
 
-public record UpkeepDto(String faction, Long numberOfArmies, Long upkeep) {
+public record UpkeepDto(String faction, Integer numberOfArmies, Integer upkeep) {
 }
